requirements.txt
Django==4.1.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
Pillow==10.2.0
python-decouple==3.8
python-barcode==0.15.1
psycopg2-binary==2.9.10 